requirements.txt
fastapi
uvicorn[standard]
pillow
ultralytics
torch
torch
torchvision
torchaudio